rename bridge/native.* to bridge/ffi_*
