+++ flutter/lib/benchmark/benchmark.dart
@@ -17,8 +17,8 @@ import 'package:wakelock/wakelock.dart';
 
 import 'package:mlcommons_ios_app/backend/bridge/handle.dart';
 import 'package:mlcommons_ios_app/backend/bridge/isolate.dart';
-import 'package:mlcommons_ios_app/backend/bridge/native.config.dart';
-import 'package:mlcommons_ios_app/backend/bridge/native.match.dart';
+import 'package:mlcommons_ios_app/backend/bridge/ffi_config.dart';
+import 'package:mlcommons_ios_app/backend/bridge/ffi_match.dart';
 import 'package:mlcommons_ios_app/backend/list.dart';
 import 'package:mlcommons_ios_app/backend/run_settings.dart';
 import 'package:mlcommons_ios_app/benchmark/benchmark_result.dart';
